Implement the logic needed to correctly handle directories

+++ src/server/pfs/server/api_server.go
@@ -4,10 +4,10 @@ import (
 	"fmt"
 	"io"
 	"math/rand"
+	"path/filepath"
 	"strings"
 	"sync"
 	"time"
-	"path/filepath"
 
 	pfsclient "github.com/pachyderm/pachyderm/src/client/pfs"
 	pfsserver "github.com/pachyderm/pachyderm/src/server/pfs"
@@ -351,6 +351,8 @@ func (a *apiServer) PutFile(putFileServer pfsclient.API_PutFileServer) (retErr e
 	var request *pfsclient.PutFileRequest
 	var err error
 	defer func(start time.Time) { a.Log(request, google_protobuf.EmptyInstance, retErr, time.Since(start)) }(time.Now())
+	a.versionLock.RLock()
+	defer a.versionLock.RUnlock()
 	ctx := versionToContext(a.version, putFileServer.Context())
 	defer func() {
 		if err := putFileServer.SendAndClose(google_protobuf.EmptyInstance); err != nil && retErr == nil {
@@ -373,6 +375,23 @@ func (a *apiServer) PutFile(putFileServer pfsclient.API_PutFileServer) (retErr e
 		return fmt.Errorf("pachyderm: leading slash in path: %s", request.File.Path)
 	}
 
+	var wg sync.WaitGroup
+	errCh := make(chan error, 1)
+	requests := []*pfsclient.PutFileRequest{request}
+	// For a file like foo/bar/buzz, we want to send two requests to create two
+	// directories foo/ and foo/bar/
+	dirs := getAncestorDirs(request.File.Path)
+	for _, dir := range dirs {
+		dirReq := &pfsclient.PutFileRequest{
+			File: &pfsclient.File{
+				Path:   dir,
+				Commit: request.File.Commit,
+			},
+			FileType: pfsclient.FileType_FILE_TYPE_DIR,
+		}
+		requests = append(requests, dirReq)
+	}
+
 	sendReq := func(request *pfsclient.PutFileRequest) (retErr error) {
 		clientConn, err := a.getClientConnForFile(request.File, a.version)
 		if err != nil {
@@ -381,6 +400,18 @@ func (a *apiServer) PutFile(putFileServer pfsclient.API_PutFileServer) (retErr e
 		putFileClient, err := pfsclient.NewInternalAPIClient(clientConn).PutFile(ctx)
 		if err != nil {
 			return err
+		}
+		if request.FileType == pfsclient.FileType_FILE_TYPE_DIR {
+			if len(request.Value) > 0 {
+				return fmt.Errorf("PutFileRequest shouldn't have type dir and a value")
+			}
+			if err := putFileClient.Send(request); err != nil {
+				return err
+			}
+			if _, err := putFileClient.CloseAndRecv(); err != nil {
+				return err
+			}
+			return nil
 		}
 		defer func() {
 			if _, err := putFileClient.CloseAndRecv(); err != nil && retErr == nil {
@@ -405,26 +436,9 @@ func (a *apiServer) PutFile(putFileServer pfsclient.API_PutFileServer) (retErr e
 		return nil
 	}
 
-	var wg sync.WaitGroup
-	errCh := make(chan error, 1)
-	requests := []*pfsclient.PutFileRequest{request}
-	// For a file like foo/bar/buzz, we want to send two requests to create two
-	// directories foo/ and foo/bar/
-	dirs := getAncestorDirs(request.File.Path)
-	for _, dir := range dirs {
-		dirReq := &pfsclient.PutFileRequest{
-			File: &pfsclient.File{
-				Path: dir,
-				Commit: request.File.Commit,
-			},
-			FileType: pfsclient.FileType_FILE_TYPE_DIR,
-		}
-		requests = append(requests, dirReq)
-	}
-
 	for _, req := range requests {
 		wg.Add(1)
-		go func() {
+		go func(req *pfsclient.PutFileRequest) {
 			defer wg.Done()
 			err := sendReq(req)
 			if err != nil {
@@ -436,7 +450,7 @@ func (a *apiServer) PutFile(putFileServer pfsclient.API_PutFileServer) (retErr e
 				}
 				return
 			}
-		}()
+		}(req)
 	}
 
 	wg.Wait()
@@ -450,9 +464,12 @@ func (a *apiServer) PutFile(putFileServer pfsclient.API_PutFileServer) (retErr e
 }
 
 func getAncestorDirs(path string) []string {
-	var ancestors  []string
-	for path != "." {
+	var ancestors []string
+	for {
 		path = filepath.Dir(path)
+		if path == "." {
+			break
+		}
 		ancestors = append(ancestors, path)
 	}
 	return ancestors
@@ -460,6 +477,8 @@ func getAncestorDirs(path string) []string {
 
 func (a *apiServer) GetFile(request *pfsclient.GetFileRequest, apiGetFileServer pfsclient.API_GetFileServer) (retErr error) {
 	defer func(start time.Time) { a.Log(request, google_protobuf.EmptyInstance, retErr, time.Since(start)) }(time.Now())
+	a.versionLock.RLock()
+	defer a.versionLock.RUnlock()
 	ctx := versionToContext(a.version, apiGetFileServer.Context())
 	clientConn, err := a.getClientConnForFile(request.File, a.version)
 	if err != nil {
@@ -481,7 +500,22 @@ func (a *apiServer) InspectFile(ctx context.Context, request *pfsclient.InspectF
 	if err != nil {
 		return nil, err
 	}
-	return pfsclient.NewInternalAPIClient(clientConn).InspectFile(ctx, request)
+
+	fileInfo, err := pfsclient.NewInternalAPIClient(clientConn).InspectFile(ctx, request)
+	if err != nil {
+		return nil, err
+	}
+
+	if fileInfo.FileType == pfsclient.FileType_FILE_TYPE_DIR {
+		// If it's a directory, chances are that its children are scattered around
+		// the cluster, so we can't get the complete list of children without
+		// doing a multicast.  So instead of returning a partial list of children,
+		// we'd rather return no children at all.  To get the complete list of
+		// children, the caller is expected to use ListFile
+		fileInfo.Children = nil
+	}
+
+	return fileInfo, nil
 }
 
 func (a *apiServer) ListFile(ctx context.Context, request *pfsclient.ListFileRequest) (response *pfsclient.FileInfos, retErr error) {
@@ -541,11 +575,52 @@ func (a *apiServer) DeleteFile(ctx context.Context, request *pfsclient.DeleteFil
 	a.versionLock.RLock()
 	defer a.versionLock.RUnlock()
 	ctx = versionToContext(a.version, ctx)
-	clientConn, err := a.getClientConnForFile(request.File, a.version)
+
+	fileInfo, err := a.InspectFile(ctx, &pfsclient.InspectFileRequest{
+		File: request.File,
+	})
 	if err != nil {
 		return nil, err
 	}
-	return pfsclient.NewInternalAPIClient(clientConn).DeleteFile(ctx, request)
+
+	if fileInfo.FileType == pfsclient.FileType_FILE_TYPE_REGULAR {
+		clientConn, err := a.getClientConnForFile(request.File, a.version)
+		if err != nil {
+			return nil, err
+		}
+		return pfsclient.NewInternalAPIClient(clientConn).DeleteFile(ctx, request)
+	}
+
+	// If we are deleting a directory, we multicast it to the entire cluster
+	clientConns, err := a.router.GetAllClientConns(a.version)
+	if err != nil {
+		return nil, err
+	}
+	var wg sync.WaitGroup
+	errCh := make(chan error, 1)
+	for _, clientConn := range clientConns {
+		wg.Add(1)
+		go func(clientConn *grpc.ClientConn) {
+			defer wg.Done()
+			_, err := pfsclient.NewInternalAPIClient(clientConn).DeleteFile(ctx, request)
+			if err != nil {
+				select {
+				case errCh <- err:
+					// error reported
+				default:
+					// not the first error
+				}
+				return
+			}
+		}(clientConn)
+	}
+	wg.Wait()
+	select {
+	case err := <-errCh:
+		return nil, err
+	default:
+	}
+	return google_protobuf.EmptyInstance, nil
 }
 
 func (a *apiServer) Version(version int64) error {
